Rename caching to cache-headers
