+++ lib/request-handler.js
@@ -11,7 +11,10 @@ const analytics = require('./analytics')
 const { makeSend } = require('./result-sender')
 const queryString = require('query-string')
 const { Inaccessible } = require('../services/errors')
-const { coalesceCacheLength, setCacheHeaders } = require('../services/caching')
+const {
+  coalesceCacheLength,
+  setCacheHeaders,
+} = require('../services/cache-headers')
 
 // We avoid calling the vendor's server for computation of the information in a
 // number of badges.
